Add fields for ACME registration: Added email address and TOS agreement fields.
These should be specified in the j8acfg.yml file in format:

connection:
  downstream:
    tls:
      acme:
        provider: letsencrypt
        email: noreply@example.org
        acceptTOS: true
        domains:
          - example.org

Probably needs extra tests, but currently works.

I've verified that it receives an error message from the ACME provider if the terms of service are not accepted,
but there is currently no validation of the email address. The email address is currently just passed to the ACME
provider for validation.